+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Generate relative file paths and GitHub URLs for Phel functions
 - Improve public Api module
+    - Add new methods and remove deprecated ones from PhelFunction
 
 ## [0.21.0](https://github.com/phel-lang/phel-lang/compare/v0.20.0...v0.21.0) - 2025-09-01